requirements.txt
fastapi
uvicorn
python-multipart
pymupdf
pytesseract
pillow
pymongo
langchain
langchain-huggingface
sentence-transformers
python-dotenv
langchain-groq
